% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_bin_sens_stan.R
\name{bin_bin_sens_stan}
\alias{bin_bin_sens_stan}
\title{Binary (M) - binary (Y) sensitivity analysis}
\usage{
bin_bin_sens_stan(outcomes, designs, prior, ...)
}
\arguments{
\item{outcomes}{List of named regression outcomes (named m and y)}

\item{designs}{List of named design matrices (named x_m, x_y, x_u)}

\item{prior}{List of named prior means and variance-covariance matrices}

\item{...}{Additional arguments passed to \code{rstan::sampling}}
}
\value{
an object of class `stanfit` returned by `rstan::sampling`
}
\description{
Binary (M) - binary (Y) sensitivity analysis
}
