% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_bin_stan.R
\name{bin_bin_stan}
\alias{bin_bin_stan}
\title{Simple binary (M) binary (Y) mediation analysis in Stan}
\usage{
bin_bin_stan(outcomes, designs, ...)
}
\arguments{
\item{outcomes}{Named list of outcomes m, u, and y}

\item{designs}{Named list of design matrices for m, u, and y}

\item{...}{Additional parameters to pass to `rstan::sampling`}
}
\value{
an object of class `stanfit` returned by `rstan::sampling`
}
\description{
Simple binary (M) binary (Y) mediation analysis in Stan
}
