% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Calculate bias of posterior mean in Stan parameter samples}
\usage{
calculate_bias(estimate, truth)
}
\arguments{
\item{estimate}{Length-P vector with each column corresponding to an estimate}

\item{truth}{Length-P vector of true parameter values}
}
\value{
Length-P vector of estimated bias
}
\description{
Calculate bias of posterior mean in Stan parameter samples
}
