% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{calculate_nder}
\alias{calculate_nder}
\title{Calculate randomized interventional analog to natural direct effect}
\usage{
calculate_nder(params = NULL, alpha = NULL, beta = NULL, gamma = NULL,
  u_ei = NULL, z1 = 0:1, z2 = 0:1)
}
\arguments{
\item{params}{List of parameters (optional)}

\item{alpha}{Vector of Y regression coefficients (optional)}

\item{beta}{Vector of M regression coefficients (optional)}

\item{gamma}{Vector of U regression coefficients (optional)}

\item{u_ei}{Whether U is exposure-induced (optional)}

\item{z1}{Vector containing levels of z1 covariate to evaluate conditional NDER}

\item{z2}{Vector containing levels of z2 covariate to evaluate conditional NDER}
}
\value{
K x 1 Matrix of conditional NDERs, for each of K covariate patterns
}
\description{
Calculate randomized interventional analog to natural direct effect
}
