% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{calculate_nder_stan}
\alias{calculate_nder_stan}
\title{Calculate NDER from Stan sensitivity analysis fit
Returns conditional on specific z1, z2 values}
\usage{
calculate_nder_stan(sens_res, ...)
}
\arguments{
\item{sens_res}{Result list with element stan_fit containing the stan fit}
}
\value{
K x R matrix of conditional NDER values for the K covariate patterns requested
from R MCMC iterations (excludes warmup)
}
\description{
Calculate NDER from Stan sensitivity analysis fit
Returns conditional on specific z1, z2 values
}
