% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{check_coverage}
\alias{check_coverage}
\title{Evaluate Y/N credible interval coverage}
\usage{
check_coverage(ci, truth)
}
\arguments{
\item{ci}{Length-2 vector or Px2 matrix of CI lower and upper bounds}

\item{truth}{Length-P vector of true parameter values}
}
\value{
Length-P logical vector of interval coverage
}
\description{
Evaluate Y/N credible interval coverage
}
