% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{get_bias_coverage}
\alias{get_bias_coverage}
\title{Calculate parameter bias and coverage from Stan model fit}
\usage{
get_bias_coverage(stan_fit, parameter_name, truth)
}
\arguments{
\item{stan_fit}{Stan fit object}

\item{parameter_name}{String name for parameter extraction}

\item{truth}{True values of parameter}
}
\value{
Length-2 list of length-P bias vector and length-P coverage
}
\description{
Calculate parameter bias and coverage from Stan model fit
}
