% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{get_bias_stan}
\alias{get_bias_stan}
\title{Calculate parameter bias and coverage from Stan model fit}
\usage{
get_bias_stan(stan_fit, pars, truth)
}
\arguments{
\item{stan_fit}{Stan fit object}

\item{pars}{Parameter names. See \code{\link[rstan]{extract}} for parsing details.}

\item{truth}{True values of parameter}

\item{parameter_name}{String name for parameter extraction}

\item{stan_fit}{Stan model fit object.}

\item{truth}{True values for parameters.}
}
\value{
Length-2 list of length-P bias, coverage, and CI width
export
Calculate bias from a stan fit

Vector of bias values.
}
\description{
Calculate parameter bias and coverage from Stan model fit
}
