% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{get_parameter_bias_coverage}
\alias{get_parameter_bias_coverage}
\title{Calculate parameter bias and coverage from a mediation fit}
\usage{
get_parameter_bias_coverage(stan_fit, params)
}
\arguments{
\item{stan_fit}{The Stan fit object (see \code{\link{run_bdf_replicate}})
Should have parameter vectors labeled alpha, beta, and gamma}

\item{params}{List of true parameter values, labeled alpha, beta, and gamma}
}
\value{
List of bias and coverage vectors for all parameters
}
\description{
Calculate parameter bias and coverage from a mediation fit
}
