% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_stan.R
\name{lm_stan}
\alias{lm_stan}
\title{Bayesian linear regression with Stan}
\usage{
lm_stan(x, y)
}
\arguments{
\item{x}{vector of inputs}

\item{y}{vector of outputs}
}
\value{
an object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian linear regression with Stan
}
