% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{make_bl_types}
\alias{make_bl_types}
\title{Construct all 4 possible combinations of binary Z1 and Z2}
\usage{
make_bl_types()
}
\value{
Named data frame of the baseline covariate types
}
\description{
Construct all 4 possible combinations of binary Z1 and Z2
}
