% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{make_bl_weights}
\alias{make_bl_weights}
\title{Calculate population prevalance of baseline covariate types}
\usage{
make_bl_weights(bl_types, pz1, pz2)
}
\arguments{
\item{bl_types}{Data frame containing the 4 unique covariate types}

\item{pz1}{P(Z1 = 1)}

\item{pz2}{P(Z2 = 1), assumed to be independent of Z1}
}
\description{
Calculate population prevalance of baseline covariate types
}
