% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{make_prior}
\alias{make_prior}
\title{Construct prior list based on type of sensitivity analysis}
\usage{
make_prior(params, prior_type, n = NULL, n_frac = 0.1)
}
\arguments{
\item{params}{List of parameters}

\item{prior_type}{Variance type: unit variance ("unit"), informative priors on
non-identifiable parameters ("partial"), strong prior information 
("strict"), or data-driven ("dd strict") based on MLE from a simulated data set}

\item{n}{Size of data set (only needed if data-driven)}

\item{n_frac}{Ratio of small n to analysis data n (only needed if data-driven)}
}
\value{
Named list of prior means and variance-covariance matrices
}
\description{
Construct prior list based on type of sensitivity analysis
}
