% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_bin_bin_sens.R
\name{make_x_m}
\alias{make_x_m}
\title{Construct design matrix for mediator regression model}
\usage{
make_x_m(z1, z2, a, u = NULL, type = c("observed", "full"))
}
\arguments{
\item{z1}{Length-N vector for baseline confounder Z1}

\item{z2}{Length-N vector for baseline confounder Z2}

\item{a}{Length-N vector for observed or hypothetical A}

\item{u}{Length-N vector for observed (true) or hypothetical U}

\item{type}{Whether to make naive matrix excluding U ("observed") or 
full matrix that includes U ("full")}
}
\value{
NxP design matrix for the M outcome model
}
\description{
Construct design matrix for mediator regression model
}
