% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data set}
\usage{
simulate_data(n, u_ei, am_intx, params = NULL)
}
\arguments{
\item{n}{Number of observations to simulate}

\item{u_ei}{0/1 flag for whether U should be exposure-induced}

\item{am_intx}{0/1 flag for A*M interaction in outcome model}

\item{params}{(Optional) List of data-generating parameters
see \code{\link{return_dgp_parameters}} for details on structure}
}
\value{
List containing: df (data frame with n rows),
outcomes (named list of m and y outcomes), and designs (named list of 
design matrices for u, m, and y regression models)
}
\description{
Simulate data set
}
