% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_bin_bin_sens.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Calculate bias of posterior mean in Stan parameter samples}
\usage{
calculate_bias(samples, truth)
}
\arguments{
\item{samples}{SxP Matrix with each column corresponding to a parameter}

\item{truth}{Length-P vector of true parameter values}
}
\value{
Length-P vector of estimated bias
}
\description{
Calculate bias of posterior mean in Stan parameter samples
}
