% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{calculate_nder_stan}
\alias{calculate_nder_stan}
\title{Calculate NDER from Stan sensitivity analysis fit
Not simulation-based! Closed form, g-formula
Returns conditional on specific z1, z2 values}
\usage{
calculate_nder_stan(stan_fit, u_ei, am_intx, ...)
}
\arguments{
\item{stan_fit}{Result list with element stan_fit containing the stan fit}

\item{u_ei}{Whether unmeasured confounder is exposure-induced}

\item{am_intx}{Whether outcome model includes exposure-mediator interaction}

\item{\dots}{Parameters to pass to \code{\link{calculate_nder}}}
}
\value{
K x R matrix of conditional NDER values for the K covariate patterns requested
from R MCMC iterations (excludes warmup)
}
\description{
Calculate NDER from Stan sensitivity analysis fit
Not simulation-based! Closed form, g-formula
Returns conditional on specific z1, z2 values
}
