% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{check_coverage}
\alias{check_coverage}
\title{Evaluate Y/N credible interval coverage for Stan samples}
\usage{
check_coverage(samples, truth)
}
\arguments{
\item{samples}{SxP Matrix with each column corresponding to a parameter}

\item{truth}{Length-P vector of true parameter values}
}
\value{
Length-P logical vector of interval coverage
}
\description{
Evaluate Y/N credible interval coverage for Stan samples
}
