% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{make_prior}
\alias{make_prior}
\title{Construct prior list based on type of sensitivity analysis}
\usage{
make_prior(alpha_mean, beta_mean, gamma_mean, prior_type)
}
\arguments{
\item{alpha_mean}{Y regression parameter mean}

\item{beta_mean}{M regression parameter mean}

\item{gamma_mean}{U regression parameter mean}

\item{prior_type}{Variance type: unit variance ("unit"), informative priors on
non-identifiable parameters ("partial"), or strong prior information 
("strict")}
}
\value{
Named list of prior means and variance-covariance matrices
}
\description{
Construct prior list based on type of sensitivity analysis
}
