% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{make_x_u}
\alias{make_x_u}
\title{Construct design matrix for unmeasured confounder regression model
Note that it does not depend on A yet!}
\usage{
make_x_u(z1, z2, a)
}
\arguments{
\item{z1}{Length-N vector for baseline confounder Z1}

\item{z2}{Length-N vector for baseline confounder Z2}

\item{a}{Length-N vector for observed or hypothetical A}
}
\value{
NxP design matrix for the U outcome model
}
\description{
Construct design matrix for unmeasured confounder regression model
Note that it does not depend on A yet!
}
