% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{return_dgp_parameters}
\alias{return_dgp_parameters}
\title{Return true DGP for binary-binary sensitivity simulations}
\usage{
return_dgp_parameters(u_ei, am_intx, yu_strength, mu_strength)
}
\arguments{
\item{u_ei}{0/1 flag for whether U should be exposure-induced}

\item{am_intx}{0/1 flag for including A*M interaction in outcome model}

\item{yu_strength}{Strength of U -> Y relationship on log-odds scale}

\item{mu_strength}{Strength of U -> M relationship on log-odds scale}
}
\value{
Named list of parameter values
}
\description{
Return true DGP for binary-binary sensitivity simulations
}
