% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_bin_bin_sens.R
\name{return_dgp_parameters}
\alias{return_dgp_parameters}
\title{Return true DGP for binary-binary sensitivity simulations
Note: does not have A -> U!}
\usage{
return_dgp_parameters()
}
\value{
Named list of parameter values
}
\description{
Return true DGP for binary-binary sensitivity simulations
Note: does not have A -> U!
}
