% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_functions.R
\name{run_sensitivity}
\alias{run_sensitivity}
\title{Run a single replicate of the binary-binary mediation sensitivity analysis}
\usage{
run_sensitivity(n, prior_type, params = NULL, ...)
}
\arguments{
\item{n}{Number of observations in data set}

\item{prior_type}{See \code{\link{make_prior}} for details}

\item{params}{List of data generating parameters (will be created by 
\code{\link{return_dgp_parameters}} if not specified)}

\item{...}{Additional parameters to be passed to bin_bin_sens_stan}
}
\value{
Stan model fit object
}
\description{
Run a single replicate of the binary-binary mediation sensitivity analysis
}
